Full base system (excluding ports) can now be compiled with clang/llvm, LibLemon: New slightly faster memcpy_optimized

+++ Kernel/src/net/net.cpp
@@ -24,7 +24,7 @@ namespace Network {
     }
 
     unsigned short AllocatePort(Socket& sock){
-        int port = EPHEMERAL_PORT_RANGE_START;
+        unsigned short port = EPHEMERAL_PORT_RANGE_START;
 
         while(port <= EPHEMERAL_PORT_RANGE_END && AcquirePort(sock, port)) port++;
